feat: Enhance Authentication & Authorization
This commit introduces several enhancements to the authentication and authorization mechanisms:

- JWTs now carry more user information, enabling richer authorization decisions and reducing database lookups.
- Authentication middleware is refined for token-based authentication, streamlining authorization across resolvers.
- Contact management resolvers now enforce authentication, ensuring only logged-in users can manage contacts.
- Client-server communication is improved by leveraging cookies for secure and convenient JWT management.

+++ src/Resolver/Contact.resolver.js
@@ -1,22 +1,45 @@
 import { Contact } from "../Models/Users/Contact.Model.js";
-const contactResolvers={
-    Query:{
-        contacts:async ()=>await Contact.find(),
-        contact:async (_,{id})=>await Contact.findById(id)
+import { AuthenticationError } from 'apollo-server-express';
+const contactResolvers = {
+    Query: {
+      contacts: async (parent, args, context) => {
+        if (!context.user) {
+          throw new AuthenticationError('You must be logged in');
+        }
+        return await Contact.find();
+      },
+      contact: async (parent, { id }, context) => {
+        console.log(context)
+        if (!context.user) {
+          throw new AuthenticationError('You must be logged in');
+        }
+        return await Contact.findById(id);
+      },
     },
-    Mutation:{
-        createContact:async (_,args)=>{
-            const newContact=new Contact(args);
-            return await newContact.save();
-        },
-        updateContact:async (_id,args)=>{
-            const updatedContact=await Contact.findByIdAndUpdate(args._id, args, {new: true});
-            return updatedContact;
-        },
-        deleteContact:async (_id, {id})=>{
-            const deletedContact=await Contact.findByIdAndDelete(id);
-            return deletedContact;
-        },
+    Mutation: {
+      createContact: async (parent, args, context) => {
+        if (!context.user) {
+          throw new AuthenticationError('You must be logged in');
+        }
+        const newContact = new Contact(args);
+        return await newContact.save();
+      },
+      updateContact: async (parent, args, context) => {
+        if (!context.user) {
+          throw new AuthenticationError('You must be logged in');
+        }
+        const updatedContact = await Contact.findByIdAndUpdate(args._id, args, { new: true });
+        return updatedContact;
+      },
+      deleteContact: async (parent, { id }, context) => {
+        if (!context.user) {
+          throw new AuthenticationError('You must be logged in');
+        }
+        const deletedContact = await Contact.findByIdAndDelete(id);
+        return deletedContact;
+      },
     },
-};
-export {contactResolvers};
+  };
+  
+  export { contactResolvers };
+
